Move app_config to config folder.
